Add "auto_reverse_comp" argument to get_seq_and_qual_barcode_read.
Add "auto_reverse_comp" argument to get_seq_and_qual_barcode_read,
so automatic reverse complimenting can be disabled if the read with
the barcode does not come from index read 2.

The SciFi-RNA barcode has a part of the barcode (and a UMI) which is
on read 1, while only the 10x ATAC barcode is on R2 and could need
reverse complementing.